Created pyserini.encode.optional; moved pillow dependency to core
