+++ pyserini/encode/__init__.py
@@ -23,6 +23,7 @@ from ._ance import AnceEncoder, AnceDocumentEncoder, AnceQueryEncoder
 from ._auto import AutoQueryEncoder, AutoDocumentEncoder
 from ._bpr import BprQueryEncoder
 from ._cached_data import CachedDataQueryEncoder
+from ._clip import ClipDocumentEncoder, ClipTextEncoder, ClipImageEncoder
 from ._cosdpr import CosDprEncoder, CosDprDocumentEncoder, CosDprQueryEncoder
 from ._dkrr import DkrrDprQueryEncoder
 from ._dpr import DprDocumentEncoder, DprQueryEncoder
