+++ pyserini/encode/__main__.py
@@ -17,11 +17,12 @@
 import argparse
 import sys
 
-from pyserini.encode import DprDocumentEncoder, TctColBertDocumentEncoder, AnceDocumentEncoder, \
-    AggretrieverDocumentEncoder, AutoDocumentEncoder, CosDprDocumentEncoder, JsonlRepresentationWriter, \
-    JsonlCollectionIterator, UniCoilDocumentEncoder, OpenAIDocumentEncoder, OPENAI_API_RETRY_DELAY
-from pyserini.encode._clip import ClipDocumentEncoder
-from pyserini.encode._faiss import FaissRepresentationWriter
+from pyserini.encode import AutoDocumentEncoder
+from pyserini.encode import AggretrieverDocumentEncoder, AnceDocumentEncoder, ClipDocumentEncoder, \
+    CosDprDocumentEncoder, DprDocumentEncoder, TctColBertDocumentEncoder, UniCoilDocumentEncoder
+from pyserini.encode import OpenAIDocumentEncoder, OPENAI_API_RETRY_DELAY
+from pyserini.encode import JsonlRepresentationWriter, JsonlCollectionIterator
+from pyserini.encode.optional import FaissRepresentationWriter
 
 encoder_class_map = {
     "dpr": DprDocumentEncoder,
