+++ pyserini/encode/optional/_faiss.py
