+++ requirements.txt
@@ -14,11 +14,11 @@ openai>=1.0.0
 sentencepiece>=0.2
 tiktoken>=0.4.0
 flask>3.0
+pillow>=10.2.0
 # ones below are going to become optional dependencies eventually:
 faiss>1.7.0
 nmslib>=2.0.6
 lightgbm>=3.3.2
 spacy>=3.2.1
 pyarrow>=15.0.0
-pillow>=10.2.0
 pybind11>=2.11.0
